pyproject.toml
[project]
name = "caption-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "koshishshresthaa", email = "koshishshrestha74@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "google-genai>=1.21.1",
    "moviepy==1.0.3",
    "openai-whisper>=20240930",
    "streamlit>=1.46.0",
    "whisperx>=3.3.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
